requirements.txt
streamlit==1.28.1
sentence-transformers==2.2.2
qdrant-client==1.7.0
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
openpyxl==3.1.2
pytest==7.4.2
scikit-learn==1.3.0
langchain==0.0.335
fpdf2==2.7.8
